fix delete the value and attribute

+++ back-end/app/Http/Controllers/Api/ItemAttributeController.php
@@ -5,10 +5,13 @@ namespace App\Http\Controllers\Api;
 
 // use App\Http\Requests\CreateAttributeRequest;
 // use App\Http\Requests\UpdateAttributeRequest;
+
+use App\Exceptions\GeneralException;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\Item\Attribute\CreateAttributeRequest;
 use App\Http\Resources\Tenant\Items\Attribute\ItemAttributeResource;
 use App\Models\Tenant\ItemAttribute;
+use DB;
 use Exception;
 use Illuminate\Http\Response;
 
@@ -41,7 +44,17 @@ class ItemAttributeController extends Controller
 
     public function destroy(ItemAttribute $attribute)
     {
-        $attribute->delete();
-        return ApiResponse(message: 'Attribute deleted successfully', code: Response::HTTP_OK);
+        try {
+            DB::beginTransaction();
+            if ($attribute->values()->exists()) {
+                throw new GeneralException(__('app.cannot_delete_item_attribute_used_by_values'));
+            }
+            $attribute->delete();
+            DB::commit();
+            return ApiResponse(message: 'Attribute deleted successfully', code: Response::HTTP_OK);
+        } catch (GeneralException $e) {
+            DB::rollBack();
+            return ApiResponse(message: $e->getMessage(), code: $e->getCode());
+        }
     }
 }
